+++ src/werkzeug/datastructures/auth_headers.py
@@ -1,102 +1,119 @@
+from __future__ import annotations
+
 import base64
+import typing as t
+import warnings
+from functools import wraps
 
-from .mixins import ImmutableDictMixin
-from .mixins import UpdateDictMixin
+from ..http import dump_header
+from ..http import parse_dict_header
+from ..http import parse_set_header
+from ..http import quote_header_value
+from .structures import CallbackDict
+from .structures import HeaderSet
+
+if t.TYPE_CHECKING:
+    import typing_extensions as te
 
 
-class Authorization(ImmutableDictMixin, dict):
-    """Represents an ``Authorization`` header sent by the client.
+class Authorization:
+    """Represents the parts of an ``Authorization`` request header.
 
-    This is returned by
-    :func:`~werkzeug.http.parse_authorization_header`. It can be useful
-    to create the object manually to pass to the test
-    :class:`~werkzeug.test.Client`.
+    :attr:`.Request.authorization` returns an instance if the header is set.
+
+    An instance can be used with the test :class:`.Client` request methods' ``auth``
+    parameter to send the header in test requests.
+
+    Depending on the auth scheme, either :attr:`parameters` or :attr:`token` will be
+    set. The ``Basic`` scheme's token is decoded into the ``username`` and ``password``
+    parameters.
+
+    For convenience, ``auth["key"]`` and ``auth.key`` both access the key in the
+    :attr:`parameters` dict, along with ``auth.get("key")`` and ``"key" in auth``.
+
+    .. versionchanged:: 2.3
+        The ``token`` parameter and attribute was added to support auth schemes that use
+        a token instead of parameters, such as ``Bearer``.
+
+    .. versionchanged:: 2.3
+        The object is no longer a ``dict``.
 
     .. versionchanged:: 0.5
-        This object became immutable.
+        The object is an immutable dict.
     """
 
-    def __init__(self, auth_type, data=None):
-        dict.__init__(self, data or {})
+    def __init__(
+        self,
+        auth_type: str,
+        data: dict[str, str] | None = None,
+        token: str | None = None,
+    ) -> None:
         self.type = auth_type
+        """The authorization scheme, like ``Basic``, ``Digest``, or ``Bearer``."""
 
-    @property
-    def username(self):
-        """The username transmitted.  This is set for both basic and digest
-        auth all the time.
+        if data is None:
+            data = {}
+
+        self.parameters = data
+        """A dict of parameters parsed from the header. Either this or :attr:`token`
+        will have a value for a give scheme.
         """
-        return self.get("username")
 
-    @property
-    def password(self):
-        """When the authentication type is basic this is the password
-        transmitted by the client, else `None`.
+        self.token = token
+        """A token parsed from the header. Either this or :attr:`parameters` will have a
+        value for a given scheme.
+
+        .. versionadded:: 2.3
         """
-        return self.get("password")
 
-    @property
-    def realm(self):
-        """This is the server realm sent back for HTTP digest auth."""
-        return self.get("realm")
+    def __getattr__(self, name: str) -> str | None:
+        return self.parameters.get(name)
 
-    @property
-    def nonce(self):
-        """The nonce the server sent for digest auth, sent back by the client.
-        A nonce should be unique for every 401 response for HTTP digest auth.
+    def __getitem__(self, name: str) -> str | None:
+        return self.parameters.get(name)
+
+    def get(self, key: str, default: str | None = None) -> str | None:
+        return self.parameters.get(key, default)
+
+    def __contains__(self, key: str) -> bool:
+        return key in self.parameters
+
+    @classmethod
+    def from_header(cls, value: str | None) -> te.Self | None:
+        """Parse an ``Authorization`` header value and return an instance, or ``None``
+        if the value is empty.
+
+        :param value: The header value to parse.
+
+        .. versionadded:: 2.3
         """
-        return self.get("nonce")
+        if not value:
+            return None
 
-    @property
-    def uri(self):
-        """The URI from Request-URI of the Request-Line; duplicated because
-        proxies are allowed to change the Request-Line in transit.  HTTP
-        digest auth only.
-        """
-        return self.get("uri")
+        scheme, _, rest = value.partition(" ")
+        scheme = scheme.lower()
+        rest = rest.strip()
 
-    @property
-    def nc(self):
-        """The nonce count value transmitted by clients if a qop-header is
-        also transmitted.  HTTP digest auth only.
-        """
-        return self.get("nc")
+        if scheme == "basic":
+            try:
+                username, _, password = base64.b64decode(rest).decode().partition(":")
+            except UnicodeError:
+                return None
 
-    @property
-    def cnonce(self):
-        """If the server sent a qop-header in the ``WWW-Authenticate``
-        header, the client has to provide this value for HTTP digest auth.
-        See the RFC for more details.
-        """
-        return self.get("cnonce")
+            return cls(scheme, {"username": username, "password": password})
 
-    @property
-    def response(self):
-        """A string of 32 hex digits computed as defined in RFC 2617, which
-        proves that the user knows a password.  Digest auth only.
-        """
-        return self.get("response")
+        parameters = parse_dict_header(rest)
 
-    @property
-    def opaque(self):
-        """The opaque header from the server returned unchanged by the client.
-        It is recommended that this string be base64 or hexadecimal data.
-        Digest auth only.
-        """
-        return self.get("opaque")
+        if len(parameters) == 1 and parameters[next(iter(parameters))] is None:
+            # There is one parameter with no value, was actually a token.
+            return cls(scheme, None, rest)
 
-    @property
-    def qop(self):
-        """Indicates what "quality of protection" the client has applied to
-        the message for HTTP digest auth. Note that this is a single token,
-        not a quoted list of alternatives as in WWW-Authenticate.
-        """
-        return self.get("qop")
+        return cls(scheme, parameters, None)
 
-    def to_header(self):
-        """Convert to a string value for an ``Authorization`` header.
+    def to_header(self) -> str:
+        """Produce an ``Authorization`` header value representing this data.
 
         .. versionadded:: 2.0
-            Added to support passing authorization to the test client.
         """
         if self.type == "basic":
             value = base64.b64encode(
@@ -104,24 +121,35 @@ class Authorization(ImmutableDictMixin, dict):
             ).decode("utf8")
             return f"Basic {value}"
 
-        if self.type == "digest":
-            return f"Digest {http.dump_header(self)}"
+        if self.token is not None:
+            return f"{self.type.title()} {self.token}"
 
-        raise ValueError(f"Unsupported type {self.type!r}.")
+        return f"{self.type.title()} {dump_header(self.parameters)}"
+
+    def __str__(self) -> str:
+        return self.to_header()
+
+    def __repr__(self) -> str:
+        return f"<{type(self).__name__} {self.to_header()}>"
 
 
-def auth_property(name, doc=None):
+def auth_property(name: str, doc: str | None = None) -> property:
     """A static helper function for Authentication subclasses to add
     extra authentication system properties onto a class::
 
         class FooAuthenticate(WWWAuthenticate):
             special_realm = auth_property('special_realm')
 
-    For more information have a look at the sourcecode to see how the
-    regular properties (:attr:`realm` etc.) are implemented.
+    .. deprecated:: 2.3
+        Will be removed in Werkzeug 2.4.
     """
+    warnings.warn(
+        "'auth_property' is deprecated and will be removed in Werkzeug 2.4.",
+        DeprecationWarning,
+        stacklevel=2,
+    )
 
-    def _set_value(self, value):
+    def _set_value(self, value):  # type: ignore[no-untyped-def]
         if value is None:
             self.pop(name, None)
         else:
@@ -130,142 +158,332 @@ def auth_property(name, doc=None):
     return property(lambda x: x.get(name), _set_value, doc=doc)
 
 
-def _set_property(name, doc=None):
-    def fget(self):
-        def on_update(header_set):
-            if not header_set and name in self:
-                del self[name]
-            elif header_set:
-                self[name] = header_set.to_header()
+class WWWAuthenticate:
+    """Represents the parts of a ``WWW-Authenticate`` response header.
 
-        return http.parse_set_header(self.get(name), on_update)
+    Set :attr:`.Response.www_authenticate` to an instance of list of instances to set
+    values for this header in the response. Modifying this instance will modify the
+    header value.
 
-    return property(fget, doc=doc)
+    Depending on the auth scheme, either :attr:`parameters` or :attr:`token` should be
+    set. The ``Basic`` scheme will encode ``username`` and ``password`` parameters to a
+    token.
 
+    For convenience, ``auth["key"]`` and ``auth.key`` both act on the :attr:`parameters`
+    dict, and can be used to get, set, or delete parameters. ``auth.get("key")`` and
+    ``"key" in auth`` are also provided.
 
-class WWWAuthenticate(UpdateDictMixin, dict):
-    """Provides simple access to `WWW-Authenticate` headers."""
+    .. versionchanged:: 2.3
+        The ``token`` parameter and attribute was added to support auth schemes that use
+        a token instead of parameters, such as ``Bearer``.
 
-    #: list of keys that require quoting in the generated header
-    _require_quoting = frozenset(["domain", "nonce", "opaque", "realm", "qop"])
+    .. versionchanged:: 2.3
+        The object is no longer a ``dict``.
 
-    def __init__(self, auth_type=None, values=None, on_update=None):
-        dict.__init__(self, values or ())
-        if auth_type:
-            self["__auth_type__"] = auth_type
-        self.on_update = on_update
+    .. versionchanged:: 2.3
+        The ``on_update`` parameter was removed.
+    """
 
-    def set_basic(self, realm="authentication required"):
-        """Clear the auth info and enable basic auth."""
-        dict.clear(self)
-        dict.update(self, {"__auth_type__": "basic", "realm": realm})
-        if self.on_update:
-            self.on_update(self)
-
-    def set_digest(
-        self, realm, nonce, qop=("auth",), opaque=None, algorithm=None, stale=False
+    def __init__(
+        self,
+        auth_type: str | None = None,
+        values: dict[str, str] | None = None,
+        token: str | None = None,
     ):
-        """Clear the auth info and enable digest auth."""
-        d = {
-            "__auth_type__": "digest",
-            "realm": realm,
-            "nonce": nonce,
-            "qop": http.dump_header(qop),
-        }
-        if stale:
-            d["stale"] = "TRUE"
-        if opaque is not None:
-            d["opaque"] = opaque
-        if algorithm is not None:
-            d["algorithm"] = algorithm
-        dict.clear(self)
-        dict.update(self, d)
-        if self.on_update:
-            self.on_update(self)
+        if auth_type is None:
+            warnings.warn(
+                "An auth type must be given as the first parameter. Assuming 'basic' is"
+                " deprecated and will be removed in Werkzeug 2.4.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+            auth_type = "basic"
 
-    def to_header(self):
-        """Convert the stored values into a WWW-Authenticate header."""
-        d = dict(self)
-        auth_type = d.pop("__auth_type__", None) or "basic"
-        kv_items = (
-            (k, http.quote_header_value(v, allow_token=k not in self._require_quoting))
-            for k, v in d.items()
+        self._type = auth_type.lower()
+        self._parameters: dict[str, str] = CallbackDict(  # type: ignore[misc]
+            values, lambda _: self._trigger_on_update()
         )
-        kv_string = ", ".join([f"{k}={v}" for k, v in kv_items])
-        return f"{auth_type.title()} {kv_string}"
+        self._token = token
+        self._on_update: t.Callable[[WWWAuthenticate], None] | None = None
 
-    def __str__(self):
-        return self.to_header()
-
-    def __repr__(self):
-        return f"<{type(self).__name__} {self.to_header()!r}>"
-
-    type = auth_property(
-        "__auth_type__",
-        doc="""The type of the auth mechanism. HTTP currently specifies
-        ``Basic`` and ``Digest``.""",
-    )
-    realm = auth_property(
-        "realm",
-        doc="""A string to be displayed to users so they know which
-        username and password to use. This string should contain at
-        least the name of the host performing the authentication and
-        might additionally indicate the collection of users who might
-        have access.""",
-    )
-    domain = _set_property(
-        "domain",
-        doc="""A list of URIs that define the protection space. If a URI
-        is an absolute path, it is relative to the canonical root URL of
-        the server being accessed.""",
-    )
-    nonce = auth_property(
-        "nonce",
-        doc="""
-        A server-specified data string which should be uniquely generated
-        each time a 401 response is made. It is recommended that this
-        string be base64 or hexadecimal data.""",
-    )
-    opaque = auth_property(
-        "opaque",
-        doc="""A string of data, specified by the server, which should
-        be returned by the client unchanged in the Authorization header
-        of subsequent requests with URIs in the same protection space.
-        It is recommended that this string be base64 or hexadecimal
-        data.""",
-    )
-    algorithm = auth_property(
-        "algorithm",
-        doc="""A string indicating a pair of algorithms used to produce
-        the digest and a checksum. If this is not present it is assumed
-        to be "MD5". If the algorithm is not understood, the challenge
-        should be ignored (and a different one used, if there is more
-        than one).""",
-    )
-    qop = _set_property(
-        "qop",
-        doc="""A set of quality-of-privacy directives such as auth and
-        auth-int.""",
-    )
+    def _trigger_on_update(self) -> None:
+        if self._on_update is not None:
+            self._on_update(self)
 
     @property
-    def stale(self):
-        """A flag, indicating that the previous request from the client
-        was rejected because the nonce value was stale.
+    def type(self) -> str:
+        """The authorization scheme, like ``Basic``, ``Digest``, or ``Bearer``."""
+        return self._type
+
+    @type.setter
+    def type(self, value: str) -> None:
+        self._type = value
+        self._trigger_on_update()
+
+    @property
+    def parameters(self) -> dict[str, str]:
+        """A dict of parameters for the header. Only one of this or :attr:`token` should
+        have a value for a give scheme.
         """
-        val = self.get("stale")
-        if val is not None:
-            return val.lower() == "true"
+        return self._parameters
+
+    @parameters.setter
+    def parameters(self, value: dict[str, str]) -> None:
+        self._parameters = CallbackDict(  # type: ignore[misc]
+            value, lambda _: self._trigger_on_update()
+        )
+        self._trigger_on_update()
+
+    @property
+    def token(self) -> str | None:
+        """A dict of parameters for the header. Only one of this or :attr:`token` should
+        have a value for a give scheme.
+        """
+        return self._token
+
+    @token.setter
+    def token(self, value: str | None) -> None:
+        """A token for the header. Only one of this or :attr:`parameters` should have a
+        value for a given scheme.
+
+        .. versionadded:: 2.3
+        """
+        self._token = value
+        self._trigger_on_update()
+
+    def set_basic(self, realm: str = "authentication required") -> None:
+        """Clear any existing data and set a ``Basic`` challenge.
+
+        .. deprecated:: 2.3
+            Will be removed in Werkzeug 2.4. Create and assign an instance instead.
+        """
+        warnings.warn(
+            "The 'set_basic' method is deprecated and will be removed in Werkzeug 2.4."
+            " Create and assign an instance instead."
+        )
+        self._type = "basic"
+        dict.clear(self.parameters)  # type: ignore[arg-type]
+        dict.update(
+            self.parameters,  # type: ignore[arg-type]
+            {"realm": realm},  # type: ignore[dict-item]
+        )
+        self._token = None
+        self._trigger_on_update()
+
+    def set_digest(
+        self,
+        realm: str,
+        nonce: str,
+        qop: t.Sequence[str] = ("auth",),
+        opaque: str | None = None,
+        algorithm: str | None = None,
+        stale: bool = False,
+    ) -> None:
+        """Clear any existing data and set a ``Digest`` challenge.
+
+        .. deprecated:: 2.3
+            Will be removed in Werkzeug 2.4. Create and assign an instance instead.
+        """
+        warnings.warn(
+            "The 'set_digest' method is deprecated and will be removed in Werkzeug 2.4."
+            " Create and assign an instance instead."
+        )
+        self._type = "digest"
+        dict.clear(self.parameters)  # type: ignore[arg-type]
+        parameters = {
+            "realm": realm,
+            "nonce": nonce,
+            "qop": ", ".join(qop),
+            "stale": "TRUE" if stale else "FALSE",
+        }
+
+        if opaque is not None:
+            parameters["opaque"] = opaque
+
+        if algorithm is not None:
+            parameters["algorithm"] = algorithm
+
+        dict.update(self.parameters, parameters)  # type: ignore[arg-type]
+        self._token = None
+        self._trigger_on_update()
+
+    def __getitem__(self, key: str) -> str | None:
+        return self.parameters.get(key)
+
+    def __setitem__(self, key: str, value: str | None) -> None:
+        if value is None:
+            if key in self.parameters:
+                del self.parameters[key]
+        else:
+            self.parameters[key] = value
+
+        self._trigger_on_update()
+
+    def __delitem__(self, key: str) -> None:
+        if key in self.parameters:
+            del self.parameters[key]
+            self._trigger_on_update()
+
+    def __getattr__(self, name: str) -> str | None:
+        return self[name]
+
+    def __setattr__(self, name: str, value: str | None) -> None:
+        if name in {"_type", "_parameters", "_token", "_on_update"}:
+            super().__setattr__(name, value)
+        else:
+            self[name] = value
+
+    def __delattr__(self, name: str) -> None:
+        del self[name]
+
+    def __contains__(self, key: str) -> bool:
+        return key in self.parameters
+
+    def get(self, key: str, default: str | None = None) -> str | None:
+        return self.parameters.get(key, default)
+
+    @classmethod
+    def from_header(cls, value: str | None) -> te.Self | None:
+        """Parse a ``WWW-Authenticate`` header value and return an instance, or ``None``
+        if the value is empty.
+
+        :param value: The header value to parse.
+
+        .. versionadded:: 2.3
+        """
+        if not value:
+            return None
+
+        scheme, _, rest = value.partition(" ")
+        scheme = scheme.lower()
+        rest = rest.strip()
+        parameters = parse_dict_header(rest)
+
+        if len(parameters) == 1 and parameters[next(iter(parameters))] is None:
+            return cls(scheme, None, rest)
+
+        return cls(scheme, parameters, None)
+
+    def to_header(self) -> str:
+        """Produce a ``WWW-Authenticate`` header value representing this data."""
+        if self.token is not None:
+            return f"{self.type.title()} {self.token}"
+
+        if self.type == "digest":
+            items = []
+
+            for key, value in self.parameters.items():
+                if key in {"realm", "domain", "nonce", "opaque", "realm", "qop"}:
+                    value = quote_header_value(value, allow_token=False)
+                else:
+                    value = quote_header_value(value)
+
+                items.append(f"{key}={value}")
+
+            return f"Digest {', '.join(items)}"
+
+        return f"{self.type.title()} {dump_header(self.parameters)}"
+
+    def __str__(self) -> str:
+        return self.to_header()
+
+    def __repr__(self) -> str:
+        return f"<{type(self).__name__} {self.to_header()}>"
+
+    @property
+    def qop(self) -> set[str]:
+        """The ``qop`` parameter as a set.
+
+        .. deprecated:: 2.3
+            Will be removed in Werkzeug 2.4. It will become the same as other
+            parameters, returning a string.
+        """
+        warnings.warn(
+            "The 'qop' property is deprecated and will be removed in Werkzeug 2.4."
+            " It will become the same as other parameters, returning a string.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+
+        def on_update(value: HeaderSet) -> None:
+            if not value:
+                if "qop" in self:
+                    del self["qop"]
+
+                return
+
+            self.parameters["qop"] = value.to_header()
+
+        return parse_set_header(self.parameters.get("qop"), on_update)
+
+    @property
+    def stale(self) -> bool | None:
+        """The ``stale`` parameter as a boolean.
+
+        .. deprecated:: 2.3
+            Will be removed in Werkzeug 2.4. It will become the same as other
+            parameters, returning a string.
+        """
+        warnings.warn(
+            "The 'stale' property is deprecated and will be removed in Werkzeug 2.4."
+            " It will become the same as other parameters, returning a string.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+
+        if "stale" in self.parameters:
+            return self.parameters["stale"].lower() == "true"
+
+        return None
 
     @stale.setter
-    def stale(self, value):
+    def stale(self, value: bool | str | None) -> None:
         if value is None:
-            self.pop("stale", None)
+            if "stale" in self.parameters:
+                del self.parameters["stale"]
+
+            return
+
+        if isinstance(value, bool):
+            warnings.warn(
+                "Setting the 'stale' property to a boolean is deprecated and will be"
+                " removed in Werkzeug 2.4.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+            self.parameters["stale"] = "TRUE" if value else "FALSE"
         else:
-            self["stale"] = "TRUE" if value else "FALSE"
+            self.parameters["stale"] = value
 
     auth_property = staticmethod(auth_property)
 
 
-# circular dependencies
-from .. import http
+def _deprecated_dict_method(f):  # type: ignore[no-untyped-def]
+    @wraps(f)
+    def wrapper(*args, **kwargs):  # type: ignore[no-untyped-def]
+        warnings.warn(
+            "Treating 'Authorization' and 'WWWAuthenticate' as a dict is deprecated and"
+            " will be removed in Werkzeug 2.4. Use the 'parameters' attribute instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        return f(*args, **kwargs)
+
+    return wrapper
+
+
+for name in (
+    "__iter__",
+    "clear",
+    "copy",
+    "items",
+    "keys",
+    "pop",
+    "popitem",
+    "setdefault",
+    "update",
+    "values",
+):
+    f = _deprecated_dict_method(getattr(dict, name))
+    setattr(Authorization, name, f)
+    setattr(WWWAuthenticate, name, f)